% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{backup}
\alias{backup}
\title{backup a file to backup/ folder, avoid overwriting it.}
\usage{
backup(filename, ifbackup = TRUE)
}
\arguments{
\item{filename}{the file to backup}

\item{ifbackup}{whether to backup}
}
\value{
backup the files
}
\description{
backup a file to backup/ folder, avoid overwriting it.
}
