% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{bd}
\alias{bd}
\title{Title Show demos}
\usage{
bd(x = template()[-which(template() == "poster")], mail_all = FALSE)
}
\arguments{
\item{x}{NA or character, templates to show}

\item{mail_all}{logical}
}
\value{
demo files
}
\description{
Title Show demos
}
\examples{
bd(template = NA)
}
