% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{bd2pd}
\alias{bd2pd}
\title{Convert a bookdown project into a page down project}
\usage{
bd2pd(proj_path = ".", book_dir = NA)
}
\arguments{
\item{proj_path}{the path of the bookdown project}

\item{book_dir}{the folder name of the rendered book files}
}
\value{
a pagedown rmd file
}
\description{
Convert a bookdown project into a page down project
}
\examples{
\dontrun{
bd2pd()
}
}
