\name{bookdownplus-imports}
\alias{image_write}
\alias{create_bib}
\alias{create_book}
\alias{create_data}
\alias{create_dir}
\alias{create_eq}
\alias{create_mindmap}
\alias{create_r}
\alias{create_rmd}
\alias{create_rosr}
\alias{create_rpkg}
\alias{create_rproj}
\alias{create_ui}
\alias{create_website}
\alias{eq}
\alias{eqw}
\alias{install_packages}
\alias{read_eq}
\alias{sub_projects}
\alias{template_ls}
\alias{templates}
\alias{download.file}
\alias{read.csv}
\alias{untar}
\alias{unzip}
\alias{inf_mr}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{magick}{\code{\link[magick:image_write]{image_write}}}
  \item{rosr}{\code{\link[rosr:create_bib]{create_bib}}, \code{\link[rosr:create_book]{create_book}}, \code{\link[rosr:create_data]{create_data}}, \code{\link[rosr:create_dir]{create_dir}}, \code{\link[rosr:create_eq]{create_eq}}, \code{\link[rosr:create_mindmap]{create_mindmap}}, \code{\link[rosr:create_r]{create_r}}, \code{\link[rosr:create_rmd]{create_rmd}}, \code{\link[rosr:create_rosr]{create_rosr}}, \code{\link[rosr:create_rpkg]{create_rpkg}}, \code{\link[rosr:create_rproj]{create_rproj}}, \code{\link[rosr:create_ui]{create_ui}}, \code{\link[rosr:create_website]{create_website}}, \code{\link[rosr:eq]{eq}}, \code{\link[rosr:eqw]{eqw}}, \code{\link[rosr:install_packages]{install_packages}}, \code{\link[rosr:read_eq]{read_eq}}, \code{\link[rosr:sub_projects]{sub_projects}}, \code{\link[rosr:template_ls]{template_ls}}, \code{\link[rosr:templates]{templates}}}
  \item{utils}{\code{\link[utils:download.file]{download.file}}, \code{\link[utils:read.csv]{read.csv}}, \code{\link[utils:untar]{untar}}, \code{\link[utils:unzip]{unzip}}}
  \item{xaringan}{\code{\link[xaringan:inf_mr]{inf_mr}}}
}}
