% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{bookdownplus}
\alias{bookdownplus}
\title{Create a demo file for an .Rmd file to compile with R bookdown package.}
\usage{
bookdownplus(template = c("article", "article_zh", "calendar", "chemistry",
  "chemistry_zh", "guitar", "mail", "journal", "yihui_mini", "yihui_demo",
  "yihui_zh", "article_mdpi", "thesis_classic", "thesis_ubt", "thesis_zju",
  "poem", "poster")[13],
  title = "R bookdown+: Authoring varied types of books and documents with R bookdown",
  author = "Peng Zhao", from_address = "Maximillian Street 27",
  from_town = "95440 Bayreuth, Germany", from_phone = "62762524",
  from_email = "dapengde@live.com", to_who = "CAMS",
  to_address = "Zhongguancunnandajie 46", to_town = "100000 Beijing, China",
  type = c("business", "personal")[2], fontsize = c("tiny", "scriptsize",
  "footnotesize", "small", "normalsize", "large", "Large", "LARGE", "huge",
  "Huge")[9], email = "pzhao@pzhao.net",
  institute = "Institute of Ecology, Univ. Innsbruck",
  longinstitute = "Institute of Ecology, University of Innsbruck, Austria",
  web = "pzhao.org", logo = "images/logo.png",
  backimg = "images/logo.png", bibliofiles = "bib/bib.bib",
  theme = c("eco", "ocean", "rose")[1])
}
\arguments{
\item{template}{character, choose a template for your book or doc}

\item{title}{poster tile}

\item{author}{poster author}

\item{from_address}{sender's address (Street and No.) for mail template}

\item{from_town}{sender's town for mail template}

\item{from_phone}{sender's phone number for mail templae}

\item{from_email}{sender's email for mail template}

\item{to_who}{receiver's name for mail template}

\item{to_address}{receiver's address (Street and No.) for mail template}

\item{to_town}{receiver's town for mail template}

\item{type}{mail type for mail template}

\item{fontsize}{fontsize of the mail body for mail template}

\item{email}{author's email for poster template}

\item{institute}{for poster template}

\item{longinstitute}{for poster template}

\item{web}{for poster template}

\item{logo}{for poster template}

\item{backimg}{for poster template}

\item{bibliofiles}{for poster template}

\item{theme}{for poster template}
}
\value{
an .Rmd file to build with bookdown
}
\description{
Yes, create a demo file for bookdown.
}
