% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{copyfolder}
\alias{copyfolder}
\title{copy necessary files to the working directory}
\usage{
copyfolder(folder = "images")
}
\arguments{
\item{folder}{the folder name}
}
\value{
copy folders to the wd.
}
\description{
copy necessary files to the working directory
}
