% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{create}
\alias{create}
\title{Prepare a strict template skeleton to contribute to bookdownplus}
\usage{
create(
  template_name = "new",
  bodyfile = "body.Rmd",
  indexfile = "index.Rmd",
  texfile = "t.tex"
)
}
\arguments{
\item{template_name}{character. tempalte name.}

\item{bodyfile}{character. name of the body file.}

\item{indexfile}{character. name of the index file.}

\item{texfile}{character. name of the texfile}
}
\value{
organzed folders and files.
}
\description{
Prepare a strict template skeleton to contribute to bookdownplus
}
\examples{
# create()
}
