% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_output}
\alias{get_output}
\title{Available output formats besides pdf_book}
\usage{
get_output(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available output formats besides pdf_book
}
\examples{
get_output()
}
