% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_template}
\alias{get_template}
\title{Available templates.}
\usage{
get_template()
}
\value{
a dataframe displaying available templates.
}
\description{
Visit \url{https://b.pzhao.org} to see a gallery of the templates.
}
\examples{
get_template()
}
