% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{include_image}
\alias{include_image}
\title{Insert an image from a url}
\usage{
include_image(img_url, img_dir = "images")
}
\arguments{
\item{img_url}{The url of the image}

\item{img_dir}{The path of the local image directory}
}
\value{
The same as the \code{ knitr::include_graphics()} function
}
\description{
This function is supposed to be used in R code chunks or inline R code expressions.
\itemize{
\item If the output is not pdf, then this function works the samge as \code{knitr::include_graphics()}.
\item If the output is pdf, then the image will be downloaded to the img_dir directory and inserted.
\item If the image is in gif format, then it will be converted into png before inserted into pdf.
}
}
