% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mail_bodysize}
\alias{mail_bodysize}
\title{Available fontsize for mail body}
\usage{
mail_bodysize(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available fontsize for mail body
}
\examples{
mail_bodysize()
}
