% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mail_font}
\alias{mail_font}
\title{Available fonts for mail body}
\usage{
mail_font(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available fonts for mail body
}
\examples{
mail_font()
}
