% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mail_fontsize}
\alias{mail_fontsize}
\title{Available fontsize for mail template}
\usage{
mail_fontsize(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available fontsize for mail template
}
\examples{
mail_fontsize()
}
