% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mail_style}
\alias{mail_style}
\title{Available styles for mail template}
\usage{
mail_style(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available styles for mail template
}
\examples{
mail_style()
}
