% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{mail_theme}
\alias{mail_theme}
\title{Available themes for mail template}
\usage{
mail_theme(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available themes for mail template
}
\examples{
mail_theme()
}
