% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{more_output}
\alias{more_output}
\title{Simple version of get_output()}
\usage{
more_output()
}
\value{
a vector of the output formats
}
\description{
Simple version of get_output()
}
\examples{
more_output()
}
