% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{poster_theme}
\alias{poster_theme}
\title{Available themes for poster templatemail body}
\usage{
poster_theme(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe.
}
\description{
Available themes for poster templatemail body
}
\examples{
poster_theme()
}
