% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{template}
\alias{template}
\title{Simple version of get_template()}
\usage{
template()
}
\value{
a vector of the template names
}
\description{
Simple version of get_template()
}
\examples{
template()
}
