% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{bd}
\alias{bd}
\title{Show demos}
\usage{
bd(template = NA, to = ".")
}
\arguments{
\item{template}{NA or character, templates to show}

\item{to}{character. The destination directory for the bookdown project}
}
\value{
demo files
}
\description{
Show demos
}
\examples{
bd(NULL)
}
