% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{bookdownplus}
\alias{bookdownplus}
\title{Create a demo file for an .Rmd file to compile with R bookdown package.}
\usage{
bookdownplus(template = "thesis_classic",
  title = "R bookdown+: Authoring varied types of books and documents with R bookdown",
  author = "Peng Zhao", from_address = "Maximillian Street 27",
  from_town = "95440 Bayreuth, Germany", from_phone = "62762524",
  from_email = "dapengde@live.com", to_who = "CAMS",
  to_address = "Zhongguancunnandajie 46", to_town = "100000 Beijing, China",
  type = c("business", "personal")[2], fontsize = c("tiny", "scriptsize",
  "footnotesize", "small", "normalsize", "large", "Large", "LARGE", "huge",
  "Huge")[9], email = "pzhao@pzhao.net",
  institute = "Institute of Ecology, Univ. Innsbruck",
  longinstitute = "Institute of Ecology, University of Innsbruck, Austria",
  web = "pzhao.org", logo = "images/logo.png",
  backimg = "images/logo.png", bibliofiles = "bib/bib.bib",
  theme = c("eco", "ocean", "rose")[1])
}
\arguments{
\item{template}{character. Choose a template for your book or doc. The default template is 'thesis_classic'. Availabel templates are:
\itemize{
\item 'article', academic article,
\item 'article_zh', academic article in Chinese,
\item 'calendar',  monthly or weekly calendar,
\item 'chemistry',   book with chemical molecular formulae and equations
\item 'chemistry_zh', chemistry book in Chinese,
\item 'guitar',  book with guitar chords,
\item 'mail', personal or business mail,
\item 'journal', laboratory journal or personal diary,
\item 'yihui_mini',  a mini demo book by Yihui Xie,
\item 'yihui_demo',  the original demo book by Yihui Xie,
\item 'yihui_zh', the demo in Chinese by Yihui Xie,
\item 'article_mdpi',   academic article by mdpi,
\item 'thesis_classic', classic thesis,
\item 'thesis_ubt',  thesis of University of Bayreuth,
\item 'thesis_zju',  thesis of Zhejiang Uni., in Chinese,
\item 'poem', poem book,
\item 'poster', poster
}}

\item{title}{character. book title.}

\item{author}{character. book author.}

\item{from_address}{character. sender's address (Street and No.) for mail template}

\item{from_town}{character. sender's town for mail template}

\item{from_phone}{character. sender's phone number for mail templae}

\item{from_email}{character. sender's email for mail template}

\item{to_who}{character. receiver's name for mail template}

\item{to_address}{character. receiver's address (Street and No.) for mail template}

\item{to_town}{character. receiver's town for mail template}

\item{type}{character. mail type for mail template}

\item{fontsize}{character. fontsize of the mail body for mail template}

\item{email}{character. author's email for poster template}

\item{institute}{character. for poster template}

\item{longinstitute}{character. for poster template}

\item{web}{character. for poster template}

\item{logo}{character. The path of the logo image. for poster template}

\item{backimg}{character. The path of the background image. for poster template}

\item{bibliofiles}{character. The path of the bib file for poster template}

\item{theme}{character. theme for poster template. available themes are 'eco' (default), 'ocean', 'rose'.}
}
\value{
an .Rmd file to build with bookdown
}
\description{
Yes, create a demo file for bookdown.
}
\examples{
bookdownplus()
}
