% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{share}
\alias{share}
\title{Prepare a template to contribute to bookdownplus}
\usage{
share(template_name = "new", bodyfile = "body.Rmd", indexfile = "index.Rmd", 
    texfile = "t.tex")
}
\arguments{
\item{template_name}{character. tempalte name.}

\item{bodyfile}{character. name of the body file.}

\item{indexfile}{character. name of the index file.}

\item{texfile}{character. name of the texfile}
}
\value{
organzed folders and files.
}
\description{
Prepare a template to contribute to bookdownplus
}
\examples{
share()
}
