% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{share}
\alias{share}
\title{Prepare a relaxed template folder to contribute to bookdownplus}
\usage{
share(template_name = NA)
}
\arguments{
\item{template_name}{character. The tempalte name.}

\item{template_path}{character. The complete path of the template folder.}
}
\value{
Prepare a folder for contribution.
}
\description{
Prepare a relaxed template folder to contribute to bookdownplus
}
\examples{
share()
}
