% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{template}
\alias{template}
\title{Available templates}
\usage{
template(dataframe = FALSE)
}
\arguments{
\item{dataframe}{logical. FALSE by default. whether return a vector or a data frame.}
}
\value{
a vector or a dataframe displaying available templates.
}
\description{
Available templates
}
\examples{
template()
}
